Add working-directory support
Building my project requires a customized ./configure step which creates
a new directory where make must be run from inside this subdirectory. As
working-directory is currently not supported for "uses:" steps, the
cleanest option to address this is in this action itself.

Adding support for providing a 'working-directory' variable allows us to
take advantage of using this scan action.